# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon2026 Backend API
  version: 1.3.0
  description: |
    OpenAPI contract for the Hackathon2026 backend.

    Authentication model:
    - `POST /auth/psychologists/login` sets `session_id` and `csrf_token` cookies.
    - `POST /auth/admin/login` sets `admin_session_id` and `csrf_token` cookies.
    - Protected `GET` endpoints require the corresponding session cookie.
    - Protected non-`GET` endpoints also require header `X-CSRF-Token` with the same value as the `csrf_token` cookie.

    Practical notes for frontend integration:
    - Base URL in local development is `http://localhost:8080`.
    - Error responses use the unified shape `{ "message": "...", "field_errors": { ... } }`.
    - Browser CORS preflight is supported for configured `ALLOWED_ORIGINS`, including credentials and the `X-CSRF-Token` header.
    - Psychologist user payloads return `portal_access_until` and `blocked_until` as `null` when they are not set.
    - Psychologist user payloads may include computed `account_status` and `subscription_status`.
    - `GET /psychologists/tests` returns activity metrics such as started/completed counts and last activity timestamps for the “Мои опросники” dashboard.
    - Test creation defaults `status` to `published` when omitted.
    - Test update defaults `status` to `draft` when omitted.
    - Public test start requires `respondent_name` and `respondent_phone`.
...
    - In-progress public sessions expire after 1 hour of inactivity.
...
servers:
...
paths:
...
  /psychologists/report-templates:
    post:
      tags:
        - Report Templates
      summary: Create report template
      description: Creates a reusable report template owned by the authenticated psychologist.
      operationId: createReportTemplate
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportTemplateInput'
      responses:
        '201':
          description: Report template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Report Templates
      summary: List report templates
      operationId: listReportTemplates
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Report templates created by the authenticated psychologist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/report-templates/{templateId}:
    parameters:
      - $ref: '#/components/parameters/TemplateIdParam'
    get:
      tags:
        - Report Templates
      summary: Get report template
      operationId: getReportTemplate
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Report template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Report Templates
      summary: Update report template
      operationId: updateReportTemplate
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReportTemplateInput'
      responses:
        '200':
          description: Updated report template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Report Templates
      summary: Delete report template
      operationId: deleteReportTemplate
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Report template deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests:
    post:
      tags:
        - Tests
      summary: Create test
      description: Creates a test owned by the authenticated psychologist.
      operationId: createPsychologistTest
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestInput'
      responses:
        '201':
          description: Test created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Tests
      summary: List tests
      description: Returns psychologist tests together with dashboard metrics used by the “Мои опросники” screen.
      operationId: listPsychologistTests
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Tests created by the authenticated psychologist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Test'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /psychologists/tests/{id}:
    parameters:
      - $ref: '#/components/parameters/TestIdParam'
    get:
      tags:
        - Tests
      summary: Get test
      operationId: getPsychologistTest
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Test details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Tests
      summary: Update test
      operationId: updatePsychologistTest
      security:
        - psychologistSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTestInput'
      responses:
        '200':
          description: Updated test
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Test'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Tests
      summary: Delete test
      operationId: deletePsychologistTest
      security:
        - psychologistSession: []
          csrfToken: []
      responses:
        '200':
          description: Test deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /psychologists/results/{sessionId}/report:
    parameters:
      - $ref: '#/components/parameters/SubmissionSessionIdParam'
      - $ref: '#/components/parameters/ReportFormatParam'
      - $ref: '#/components/parameters/ReportAudienceParam'
    get:
      tags:
        - Tests
      summary: Generate report for psychologist or client
      description: |
        Generates a report on demand for a completed submission.
        The report is never stored as a file on the server and is rebuilt for each request.
        By default `audience=psychologist` returns a technical report for the specialist.
        Use `audience=client` to generate the detailed client-facing report that the psychologist can later send manually to the client.
        Use `format=html` or `format=docx`.
      operationId: getPsychologistReportBySessionId
      security:
        - psychologistSession: []
      responses:
        '200':
          description: Generated psychologist report
          content:
            text/html:
              schema:
                type: string
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    psychologistSession:
      type: apiKey
      in: cookie
      name: session_id
      description: Psychologist session cookie returned by the psychologist login endpoint.
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session_id
      description: Admin session cookie returned by the admin login endpoint.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: "Must match the `csrf_token` cookie value on protected non-GET requests."
  parameters:
    TestIdParam:
      name: id
      in: path
      required: true
      description: Test identifier
      schema:
        type: integer
        format: int64
        minimum: 1
...
    SubmissionSessionIdParam:
      name: sessionId
      in: path
      required: true
      description: Public test session identifier
      schema:
        type: integer
        format: int64
        minimum: 1
    TemplateIdParam:
      name: templateId
      in: path
      required: true
      description: Report template identifier
      schema:
        type: integer
        format: int64
        minimum: 1
...
    ReportFormatParam:
      name: format
      in: query
      required: false
      description: Report output format. Defaults to `html`.
      schema:
        type: string
        enum:
          - html
          - docx
    ReportAudienceParam:
      name: audience
      in: query
      required: false
      description: Report audience. Defaults to `psychologist`. Use `client` only from the psychologist workspace when preparing the report for later delivery to the client.
      schema:
        type: string
        enum:
          - psychologist
          - client
  responses:
    BadRequest:
      description: Invalid request payload or path parameter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired session credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated request is not allowed or fails CSRF/origin checks.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
        - field_errors
      properties:
        message:
          type: string
          example: Validation failed
        field_errors:
          type: object
          nullable: true
          additionalProperties:
            type: string
          example:
            email: Email already exists
    StatusResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
...
    ReportTemplate:
      type: object
      required:
        - id
        - name
        - description
        - template_body
        - created_by
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        template_body:
          type: string
          description: |
            JSON configuration stored as a string. It supports separate `client` and `psychologist`
            sections with optional `title`, `meta_labels`, `section_titles`, `intro_paragraphs`,
            `closing_paragraphs`, and `chart_caption`.
          example: |
            {
              "client": {
                "title": "Индивидуальный отчет ПрофДНК",
                "section_titles": {
                  "summary": "Короткий вывод",
                  "chart_data": "График профиля"
                }
              },
              "psychologist": {
                "title": "Технический отчет профориентолога",
                "section_titles": {
                  "raw_scores": "Сырые показатели",
                  "answers_table": "Таблица ответов"
                }
              }
            }
        created_by:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateReportTemplateInput:
      type: object
      additionalProperties: false
      required:
        - name
        - template_body
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        template_body:
          type: string
          minLength: 2
          description: JSON configuration string for the report template.
    UpdateReportTemplateInput:
      type: object
      additionalProperties: false
      required:
        - name
        - template_body
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        template_body:
          type: string
          minLength: 2
          description: JSON configuration string for the report template.
    Test:
      type: object
      required:
        - id
        - title
        - description
        - created_by_user_id
        - has_participant_limit
        - status
        - is_public
        - started_sessions_count
        - in_progress_sessions_count
        - completed_sessions_count
        - last_started_at
        - last_completed_at
        - last_activity_at
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        created_by_user_id:
          type: integer
          format: int64
        report_template_id:
          type: integer
          format: int64
        recommended_duration:
          type: integer
          description: Recommended duration in minutes.
          minimum: 0
        max_participants:
          type: integer
          description: Maximum number of respondents allowed via the public link when `has_participant_limit` is `true`.
          minimum: 0
        has_participant_limit:
          type: boolean
          description: Whether the public link is restricted to a fixed number of respondents.
        collect_respondent_age:
          type: boolean
        collect_respondent_gender:
          type: boolean
        collect_respondent_education:
          type: boolean
        status:
          type: string
          enum:
            - draft
            - published
        public_slug:
          type: string
        is_public:
          type: boolean
        public_url:
          type: string
        started_sessions_count:
          type: integer
          minimum: 0
          description: Total number of started public sessions for the test.
        in_progress_sessions_count:
          type: integer
          minimum: 0
          description: Number of sessions currently in progress.
        completed_sessions_count:
          type: integer
          minimum: 0
        last_started_at:
          type: string
          format: date-time
          nullable: true
        last_completed_at:
          type: string
          format: date-time
          nullable: true
        last_activity_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTestInput:
      type: object
      additionalProperties: false
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        report_template_id:
          type: integer
          format: int64
        recommended_duration:
          type: integer
          minimum: 0
        max_participants:
          type: integer
          description: Positive respondent limit used when `has_participant_limit` is `true`. Send `0` to keep the link available to everyone.
          minimum: 0
        has_participant_limit:
          type: boolean
          description: When `true`, `max_participants` must be greater than `0`. When omitted, the backend infers the value from `max_participants` for backward compatibility.
        collect_respondent_age:
          type: boolean
        collect_respondent_gender:
          type: boolean
        collect_respondent_education:
          type: boolean
        status:
          type: string
          description: "Defaults to `published` when omitted or empty."
          enum:
            - draft
            - published
          default: published
    UpdateTestInput:
      type: object
      additionalProperties: false
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        report_template_id:
          type: integer
          format: int64
        recommended_duration:
          type: integer
          minimum: 0
        max_participants:
          type: integer
          description: Positive respondent limit used when `has_participant_limit` is `true`. Send `0` to keep the link available to everyone.
          minimum: 0
        has_participant_limit:
          type: boolean
          description: When `true`, `max_participants` must be greater than `0`. When omitted, the backend infers the value from `max_participants` for backward compatibility.
        collect_respondent_age:
          type: boolean
        collect_respondent_gender:
          type: boolean
        collect_respondent_education:
          type: boolean
        status:
          type: string
          description: "Defaults to `draft` when omitted or empty."
          enum:
            - draft
            - published
          default: draft
...